refactor(api): migrate usecase find-pending-organization-invitations
