+++ api/tests/team/unit/domain/usecases/find-pending-organization-invitations_test.js
@@ -1,9 +1,9 @@
-import { findPendingOrganizationInvitations } from '../../../../lib/domain/usecases/find-pending-organization-invitations.js';
-import { OrganizationInvitation } from '../../../../src/team/domain/models/OrganizationInvitation.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { OrganizationInvitation } from '../../../../../src/team/domain/models/OrganizationInvitation.js';
+import { findPendingOrganizationInvitations } from '../../../../../src/team/domain/usecases/find-pending-organization-invitations.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Unit | UseCase | find-pending-organization-invitations', function () {
-  it('should succeed', async function () {
+describe('Unit | Team | UseCase | find-pending-organization-invitations', function () {
+  it('succeeds', async function () {
     // given
     const organizationId = 1234;
     const organizationInvitationRepositoryStub = { findPendingByOrganizationId: sinon.stub() };
@@ -21,7 +21,7 @@ describe('Unit | UseCase | find-pending-organization-invitations', function () {
     });
   });
 
-  it('should return the OrganizationInvitations belonging to the given organization', async function () {
+  it('returns the OrganizationInvitations belonging to the given organization', async function () {
     // given
     const organizationId = 1234;
     const foundOrganizationInvitations = [
